added put to add update account functionality

+++ bankapp/src/main/java/com/banking/bankapp/controller/BankAccountController.java
@@ -27,17 +27,37 @@ public class BankAccountController {
         else return ResponseEntity.status(200).body(account);
     }
     @PostMapping()
-    public BankAccount saveAccount(@RequestBody BankAccount bankAccount){
-        return bankAccountService.createBankAccount(bankAccount);
+    public ResponseEntity<?> saveAccount(@RequestBody BankAccount bankAccount){
+        BankAccount account = bankAccountService.createBankAccount(bankAccount);
+        if(account != null){
+            System.out.println("Account created");
+            return ResponseEntity.status(201).body(account);}
+        return ResponseEntity.status(404).body("Something went wrong");
+
     }
     @DeleteMapping("/{id}")
-    public String deleteAccountById(@PathVariable Long id){
-         bankAccountService.deleteBankAccount(id);
-         return "Deleted account with id: "+id;
+    public ResponseEntity<?> deleteAccountById(@PathVariable Long id){
+        BankAccount account = bankAccountService.findBankAccountByID(id);
+        if(account != null){
+            bankAccountService.deleteBankAccount(id);
+            return ResponseEntity.status(200).body("Account deleted with ID: "+id+" successfully");
+        }
+        return ResponseEntity.status(404).body("Account doesn't exist");
     }
 
     @GetMapping("/test")
     public String test(){
         return "Test is working";
+    }
+
+    @PutMapping("/{id}")
+    public ResponseEntity<?> updateAccount(@PathVariable long id,@RequestBody BankAccount account){
+        BankAccount existingAccount = bankAccountService.findBankAccountByID(id);
+        if(existingAccount == null){
+            return ResponseEntity.status(404).body("no account exists with Id: "+id);
+        }
+        bankAccountService.updateAccount(id,account);
+        return ResponseEntity.status(200).body("Account with Id "+id+" is updated");
+
     }
 }
